+++ ingest/jobs/Get_Ingest_Config/tests/unit/test_get_ingest_config.py
@@ -2,32 +2,40 @@ import logging
 
 from collections import Counter
 from jsonschema import validate
+from jsonschema.exceptions import ValidationError, SchemaError
 
 from utils.config_utils import load_config_as_dict, load_configs_as_list
 from utils.constants import INGEST_SOURCES_DIR, INGEST_SOURCES_SCHEMA_PATH
+from utils.testing_utils import config_uniqueness_check
 
 
 def test_config_ingest_sources_schema_valid():
     schema = load_config_as_dict(INGEST_SOURCES_SCHEMA_PATH, "ingest")
     all_configs = load_configs_as_list(INGEST_SOURCES_DIR, "ingest")
+    errors = []
     for config in all_configs:
-        validate(config, schema)
+        try:
+            validate(config, schema)
+        except Exception as e:
+            errors.append(e)
+
+    if errors:
+        raise Exception(errors)
 
 
 def test_config_ingest_sources_uniqueness():
     all_configs = load_configs_as_list(INGEST_SOURCES_DIR, "ingest")
-    assert config_uniqueness_check(all_configs, "data_source_name")
+    errors = []
+    try:
+        assert config_uniqueness_check(all_configs, "data_source_name")
+    except Exception as e:
+        errors.append(e)    
     for config in all_configs:
-        ingest_entities = config["ingest_entities"]
-        assert config_uniqueness_check(ingest_entities, "display_name")
+        try:
+            ingest_entities = config["ingest_entities"]
+            assert config_uniqueness_check(ingest_entities, "display_name")
+        except Exception as e:
+            errors.append(e)
 
-
-def config_uniqueness_check(config_list: list[dict], unique_key: str):
-    key_counter = Counter(i[unique_key] for i in config_list)
-    for config in config_list:
-        if key_counter[config[unique_key]] == 1:
-            continue
-        else:
-            logging.error(f"{unique_key} value is not unique: {config[unique_key]}")
-            return False
-    return True
+    if errors:
+        raise Exception(errors)
