+++ utils/test/unit/test_utils.py
@@ -5,6 +5,7 @@ from jsonschema import validate
 
 from utils.config_utils import load_config_as_dict, load_configs_as_list
 from utils.constants import INGEST_SOURCES_DIR, INGEST_SOURCES_SCHEMA_PATH
+from utils.testing_utils import config_uniqueness_check
 
 
 def test_config_ingest_sources_schema_valid():
@@ -16,18 +17,17 @@ def test_config_ingest_sources_schema_valid():
 
 def test_config_ingest_sources_uniqueness():
     all_configs = load_configs_as_list(INGEST_SOURCES_DIR, "utils/test")
-    assert config_uniqueness_check(all_configs, "data_source_name")
+    errors = []
+    try:
+        assert config_uniqueness_check(all_configs, "data_source_name")
+    except Exception as e:
+        errors.append(e)    
     for config in all_configs:
-        ingest_entities = config["ingest_entities"]
-        assert config_uniqueness_check(ingest_entities, "display_name")
+        try:
+            ingest_entities = config["ingest_entities"]
+            assert config_uniqueness_check(ingest_entities, "display_name")
+        except Exception as e:
+            errors.append(e)
 
-
-def config_uniqueness_check(config_list: list[dict], unique_key: str):
-    key_counter = Counter(i[unique_key] for i in config_list)
-    for config in config_list:
-        if key_counter[config[unique_key]] == 1:
-            continue
-        else:
-            logging.error(f"{unique_key} value is not unique: {config[unique_key]}")
-            return False
-    return True
+    if errors:
+        raise Exception(errors)
